Add DummyService and update styles
