+++ App/src/main/java/com/gluonhq/chat/GluonChat.java
@@ -1,13 +1,17 @@
 package com.gluonhq.chat;
 
+import com.airhacks.afterburner.injection.Injector;
 import com.gluonhq.attach.lifecycle.LifecycleService;
+import com.gluonhq.charm.glisten.afterburner.GluonInstanceProvider;
 import com.gluonhq.charm.glisten.application.MobileApplication;
 import com.gluonhq.charm.glisten.visual.Swatch;
+import com.gluonhq.chat.service.CloudlinkService;
+import com.gluonhq.chat.service.DummyService;
+import com.gluonhq.chat.service.Service;
 import com.gluonhq.chat.views.AppViewManager;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
-import org.scenicview.ScenicView;
 
 import java.io.IOException;
 import java.util.logging.Level;
@@ -25,6 +29,12 @@ public class GluonChat extends MobileApplication {
             LOG.log(Level.SEVERE, "Storage Service Error", e);
         }
     }
+
+    private static final GluonInstanceProvider instanceSupplier = new GluonInstanceProvider() {{
+        bindProvider(Service.class, CloudlinkService::new);
+
+        Injector.setInstanceSupplier(this);
+    }};
 
     @Override
     public void init() {
@@ -69,7 +79,7 @@ public class GluonChat extends MobileApplication {
         scene.getWindow().setOnCloseRequest(e ->
                 LifecycleService.create().ifPresent(LifecycleService::shutdown));
 
-        ScenicView.show(scene);
+        //ScenicView.show(scene);
     }
 
     public static void main(String[] args) {
